+++ lib/src/config.defines.dart
@@ -1,12 +1,16 @@
 // Flutter imports:
 import 'package:flutter/cupertino.dart';
+
 // Package imports:
 import 'package:zego_uikit/zego_uikit.dart';
 
-typedef ZegoCallAudioVideoContainerBuilder = Widget Function(
-  BuildContext,
+typedef ZegoCallAudioVideoContainerBuilder = Widget? Function(
+  BuildContext context,
   List<ZegoUIKitUser> allUsers,
   List<ZegoUIKitUser> audioVideoUsers,
+
+  /// The default audio-video view creator, you can also use [ZegoAudioVideoView] as a child control to continue encapsulating
+  ZegoAudioVideoView Function(ZegoUIKitUser) audioVideoViewCreator,
 );
 
 ///  the configuration for the hang-up confirmation dialog
@@ -31,4 +35,3 @@ class ZegoCallHangUpConfirmDialogInfo {
     this.confirmButtonName = 'OK',
   });
 }
-