+++ lib/src/invitation/service.dart
@@ -216,7 +216,7 @@ class ZegoUIKitPrebuiltCallInvitationService
 
     await ZegoUIKit().getZegoUIKitVersion().then((uikitVersion) {
       ZegoLoggerService.logInfo(
-        'versions: zego_uikit_prebuilt_call:4.7.3; $uikitVersion',
+        'versions: zego_uikit_prebuilt_call:4.8.0; $uikitVersion',
         tag: 'call',
         subTag: 'call invitation service(${identityHashCode(this)})',
       );