+++ lib/src/minimizing/data.dart
@@ -5,11 +5,14 @@ import 'dart:core';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-// Project imports:
-
+// Package imports:
 import 'package:zego_uikit/zego_uikit.dart';
+
+// Project imports:
 import 'package:zego_uikit_prebuilt_call/src/config.dart';
 import 'package:zego_uikit_prebuilt_call/src/events.dart';
+
+// Project imports:
 
 /// @nodoc
 class ZegoCallMinimizeData {